Restructure the app's folders and files
